feat(auth): handle logout differently for Token and Bearer auth

+++ src/controllers/auth.js
@@ -89,8 +89,6 @@ module.exports = {
     });
   },
 
-  // CRUD:
-
   logout: async (req, res) => {
     /*
                 #swagger.tags = ["Tokens"]
@@ -99,12 +97,19 @@ module.exports = {
 
     const auth = req.headers?.authorization; //"Token token"
     const tokenKey = auth ? auth.split(" ") : null; // [ "Token", tokenKey]
-    const result = await Token.deleteOne({ token: tokenKey[1] });
+    if (tokenKey[0] == "Token") {
+      const result = await Token.deleteOne({ token: tokenKey[1] });
 
-    res.send({
-      error: false,
-      message: "Token deleted. Logout was OK.",
-      result,
-    });
+      res.send({
+        error: false,
+        message: "Token deleted. Logout was OK.",
+        result,
+      });
+    } else if (tokenKey[0] == "Bearer") {
+      res.send({
+        error: false,
+        message: "JWT: No need any process for logout.",
+      });
+    }
   },
 };
